requirements.txt
flask==3.0.0
twilio==8.11.0
python-dotenv==1.0.0
google-generativeai==0.8.3
pytest==8.3.3
pytest-cov==4.1.0
